omp: introducing requires_threads_sync in some kernel args to avoid hanging threads
We can't have threads exits early in openmp and use barrier in the rest of the threads.
OpenMP requires all the threads in a team call participate in a barrier.
In order to help the kernel launcher to filter out potentially hanging kernels,
some kernel args gets a static constexpr ThreadsSync requires_threads_sync,
which tells the kernel launcher if any of the x/y/z thread blocks has to be
divisible to the total threads.